Upgrade to KafkaFlow v3.0
