+++ samples/KafkaFlow.Serializer.MessagePack.Sample/KafkaFlow.Serializer.MessagePack.Sample.csproj
@@ -2,7 +2,7 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <IsPackable>false</IsPackable>
         <GenerateDocumentationFile>false</GenerateDocumentationFile>
         <LangVersion>9</LangVersion>
@@ -17,11 +17,10 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="KafkaFlow" Version="2.2.13" />
-        <PackageReference Include="KafkaFlow.LogHandler.Console" Version="2.2.13" />
-        <PackageReference Include="KafkaFlow.Microsoft.DependencyInjection" Version="2.2.13" />
-        <PackageReference Include="KafkaFlow.Serializer" Version="2.2.13" />
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.17" />
+        <PackageReference Include="KafkaFlow" Version="3.0.1" />
+        <PackageReference Include="KafkaFlow.LogHandler.Console" Version="3.0.1" />
+        <PackageReference Include="KafkaFlow.Microsoft.DependencyInjection" Version="3.0.1" />
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.1" />
         <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
     </ItemGroup>
