+++ src/KafkaFlow.UnitTests/MessagePackSerializerTests.cs
@@ -6,7 +6,6 @@ namespace KafkaFlow.Serializer.MessagePack.UnitTests
     using global::MessagePack.Resolvers;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Moq;
-    using System;
     using System.IO;
     using System.Threading.Tasks;
 
@@ -39,39 +38,6 @@ namespace KafkaFlow.Serializer.MessagePack.UnitTests
             // Assert
             output.Length.Should().BeGreaterThan(0);
             output.Position.Should().BeGreaterThan(0);
-        }
-
-        [TestMethod]
-        public async Task DeserializeAsync_ValidPayload_ObjectGenerated()
-        {
-            // Arrange
-            var message = fixture.Create<TestMessage>();
-            using var input = new MemoryStream();
-            await MessagePackSerializer.SerializeAsync(message.GetType(), input, message, MessagePackOptions);
-            input.Seek(0, SeekOrigin.Begin);
-
-            // Act
-            var result = await serializer.DeserializeAsync(input, typeof(TestMessage), contextMock.Object);
-
-            // Assert
-            result.Should().NotBeNull();
-            result.Should().BeOfType<TestMessage>();
-        }
-
-        [MessagePackObject]
-        private class TestMessage
-        {
-            [Key(0)]
-            public int IntegerField { get; set; }
-
-            [Key(1)]
-            public string StringField { get; set; }
-
-            [Key(2)]
-            public double DoubleField { get; set; }
-
-            [Key(3)]
-            public DateTime DateTimeField { get; set; }
         }
     }
 }
